+++ contracts/interfaces/klayswap/IFactory.sol
@@ -2,6 +2,19 @@
 pragma solidity ^0.8.9;
 
 interface IFactory {
-    function createKctPool(address tokenA, uint amountA, address tokenB, uint amountB, uint fee) external;
-    function exchangeKctPos(address tokenA, uint amountA, address tokenB, uint amountB, address[] memory path) external;
+    function createKctPool(
+        address tokenA,
+        uint256 amountA,
+        address tokenB,
+        uint256 amountB,
+        uint256 fee
+    ) external;
+
+    function exchangeKctPos(
+        address tokenA,
+        uint256 amountA,
+        address tokenB,
+        uint256 amountB,
+        address[] memory path
+    ) external;
 }
